+++ src/components/SelectImageListRow/SelectImageListRow.scss
@@ -14,7 +14,6 @@
   color: rgba(0, 0, 0, 0.87);
 }
 .select-platform span {
-  margin-bottom: 1vh;
   font-family: "Pretendard";
   font-style: normal;
   font-weight: 400;
@@ -26,8 +25,6 @@
 .select-platform ul {
   display: flex;
   flex-direction: row;
-  margin-left: 0;
-  padding-left: 0;
   list-style-type: none;
 }
 .select-platform ul li {
@@ -44,4 +41,8 @@
   line-height: 20px;
 
   color: rgba(0, 0, 0, 0.87);
+}
+.select-platform .unselect-items {
+  width: 75%;
+  height: 7vw;
 }
